Extract `types` code into separate file
