+++ internal/bson/array.go
@@ -17,12 +17,9 @@ package bson
 import (
 	"bytes"
 	"encoding/binary"
-	"errors"
 	"log/slog"
 	"strconv"
 
-	"github.com/FerretDB/FerretDB/internal/types"
-	"github.com/FerretDB/FerretDB/internal/util/iterator"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/util/must"
 )
@@ -55,34 +52,6 @@ func MakeArray(cap int) *Array {
 	}
 }
 
-// ConvertArray converts [*types.Array] to Array.
-func ConvertArray(arr *types.Array) (*Array, error) {
-	iter := arr.Iterator()
-	defer iter.Close()
-
-	elements := make([]any, arr.Len())
-
-	for {
-		i, v, err := iter.Next()
-		if err != nil {
-			if errors.Is(err, iterator.ErrIteratorDone) {
-				return &Array{
-					elements: elements,
-				}, nil
-			}
-
-			return nil, lazyerrors.Error(err)
-		}
-
-		v, err = convertFromTypes(v)
-		if err != nil {
-			return nil, lazyerrors.Error(err)
-		}
-
-		elements[i] = v
-	}
-}
-
 // Freeze prevents array from further modifications.
 // Any methods that would modify the array will panic.
 //
@@ -96,27 +65,6 @@ func (arr *Array) checkFrozen() {
 	if arr.frozen {
 		panic("array is frozen and can't be modified")
 	}
-}
-
-// Convert converts Array to [*types.Array], decoding raw documents and arrays on the fly.
-func (arr *Array) Convert() (*types.Array, error) {
-	values := make([]any, len(arr.elements))
-
-	for i, f := range arr.elements {
-		v, err := convertToTypes(f)
-		if err != nil {
-			return nil, lazyerrors.Error(err)
-		}
-
-		values[i] = v
-	}
-
-	res, err := types.NewArray(values...)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	return res, nil
 }
 
 // Len returns the number of elements in the Array.
