+++ internal/bson/bson.go
@@ -292,65 +292,3 @@ func convertToTypes(v any) (any, error) {
 		panic(fmt.Sprintf("invalid BSON type %T", v))
 	}
 }
-
-// Convert converts valid types package values to BSON values of that package.
-//
-// Conversions of composite types may cause errors.
-func Convert[T types.Type](v T) (any, error) {
-	return convertFromTypes(v)
-}
-
-// convertFromTypes is a variant of [Convert] without type parameters (generics).
-//
-// Invalid types cause panics.
-func convertFromTypes(v any) (any, error) {
-	switch v := v.(type) {
-	case *types.Document:
-		doc, err := ConvertDocument(v)
-		if err != nil {
-			return nil, lazyerrors.Error(err)
-		}
-
-		return doc, nil
-
-	case *types.Array:
-		arr, err := ConvertArray(v)
-		if err != nil {
-			return nil, lazyerrors.Error(err)
-		}
-
-		return arr, nil
-
-	case float64:
-		return v, nil
-	case string:
-		return v, nil
-	case types.Binary:
-		return Binary{
-			B:       v.B,
-			Subtype: BinarySubtype(v.Subtype),
-		}, nil
-	case types.ObjectID:
-		return ObjectID(v), nil
-	case bool:
-		return v, nil
-	case time.Time:
-		return v, nil
-	case types.NullType:
-		return Null, nil
-	case types.Regex:
-		return Regex{
-			Pattern: v.Pattern,
-			Options: v.Options,
-		}, nil
-	case int32:
-		return v, nil
-	case types.Timestamp:
-		return Timestamp(v), nil
-	case int64:
-		return v, nil
-
-	default:
-		panic(fmt.Sprintf("invalid type %T", v))
-	}
-}
